ch - refactored my-courses page component and created new component
