+++ src/Pages/My-Courses-Page/My-Courses-Page.component.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Route } from "react-router-dom";
+
+import MyCourses from '../../components/My-Courses-Overview/My-Courses-Overview.component';
+
+const MyCoursesPage = ({ match }) => {
+  console.log(match);
+  return (
+    <div>
+      <Route exact path={`${match.path}`} component={MyCourses} />
+    </div>
+  );
+};
+
+export default MyCoursesPage;
