+++ src/components/My-Courses-Overview/My-Courses-Overview.styles.jsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const MyCoursesPageContainer = styled.div`
+  font-family: "Poppins", sans-serif;
+`;
+
+export const MyCoursesContainer = styled.div`
+  margin: 2em 6.5vw 4em;
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-gap: 2em;
+`;
